python/requirements.txt
mysql-connector-python
matplotlib
pyyaml
